Swap paths url options + move paths cache files + allow execution from anywhere
The old and new paths urls were swapped when they are tried to avoid
the warnings in log when trying the old url first.

Cache files are now stored and loaded from the cache dir root, the
subdirs were dropped from their path because it didn't make much
sense to use them and their names were very long and full of special
chars.

The minimize-json.sh script is now executed with absolute path.